# cranelift-wasm: Only allocate if vectors need bitcasts